working build of split files
